pyproject.toml
[project]
name = "fastapi-boilerplate"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.17.2",
    "asyncpg>=0.31.0",
    "fastapi>=0.115.0",
    "fastapi-pagination[sqlalchemy]>=0.15.3",
    "fastapi-users[sqlalchemy]>=15.0.1",
    "httpx-oauth>=0.16.1",
    "loguru>=0.7.3",
    "pydantic-settings>=2.12.0",
    "tenacity>=9.0.0",
    "redis[hiredis]>=7.1.0",
    "sqlmodel>=0.0.27",
    "uvicorn[standard]>=0.32.0",
]

[dependency-groups]
dev = [
    "fakeredis>=2.32.1",
    "pytest>=8.3.3",
    "pytest-asyncio>=1.2.0",
    "pytest-cov>=6.1.1",
    "pytest-timeout>=2.4.0",
    "python-dotenv>=1.2.1",
    "ruff>=0.12.8",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
# filterwarnings = ["error::DeprecationWarning"]
timeout = 5
pythonpath = ["."]
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests",
]

[[tool.uv.index]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"
